examples/AzureMLSeedDet/requirements.txt
azureml-inference-server-http
